clean compile

+++ fm-domain/src/main/java/io/zeitwert/fm/account/service/api/impl/AccountServiceImpl.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.jooq.DSLContext;
 import org.springframework.stereotype.Service;
 
-import io.dddrive.oe.model.ObjTenant;
+import io.dddrive.core.oe.model.ObjTenant;
 import io.zeitwert.fm.account.model.db.Tables;
 import io.zeitwert.fm.account.model.db.tables.records.ObjAccountVRecord;
 import io.zeitwert.fm.account.service.api.AccountService;
@@ -23,7 +23,7 @@ public class AccountServiceImpl implements AccountService {
 	public List<ObjAccountVRecord> getAccounts(ObjTenant tenant) {
 		return this.dslContext
 				.selectFrom(Tables.OBJ_ACCOUNT_V)
-				.where(Tables.OBJ_ACCOUNT_V.TENANT_ID.eq(tenant.getId()))
+				.where(Tables.OBJ_ACCOUNT_V.TENANT_ID.eq((Integer) tenant.getId()))
 				.fetch();
 	}
